Progress on Tables
Initial coding for generating the MAG coverage table and
made a start on the "value add" table. Will require a bit
of cleanup but looks like it should work to replicate
Bianca's current versions reasonably well.

TODO - the version with the crossref types. Should be a
straightforward GroupBy and re-use of the convenience function